Authorization added for all transactions

+++ controller/transaction.js
@@ -1,8 +1,11 @@
 const Transaction = require('../Models/Transaction')
+const User = require('../Models/User')
+
+
 
 exports.getTransactions = async (req,res,next) => {
     try {
-        const transactions= await Transaction.find();
+        const transactions= await Transaction.find({user: req.user.id});
 
         return res.status(200).json({
             success: true,
@@ -19,7 +22,11 @@ exports.getTransactions = async (req,res,next) => {
 
 exports.addTransactions = async (req,res,next) => {
     try {
-        const transaction = await Transaction.create(req.body);
+        const transaction = await Transaction.create({
+            user: req.user.id,
+            text: req.body.text,
+            amount: req.body.amount
+        });
 
         return res.status(201).json({
             data: transaction,
@@ -39,7 +46,7 @@ exports.addTransactions = async (req,res,next) => {
         {
             return res.status(500).json({
                 success: false,
-                error: `Error: ${err.message}`
+                error: `Error: ${error.message}`
             })
         }
     }
@@ -56,6 +63,17 @@ exports.deleteTransactions = async (req,res,next) => {
             })
         }
         
+        const user = await User.findById(req.user.id);
+        if(!user)
+        {
+            return res.status(401).json({error:'User not found'})
+        }
+        
+        if (transaction.user.toString() !== req.user.id) {
+            res.status(401)
+            throw new Error('User not authorized')
+          }
+
             await transaction.deleteOne();
             return res.status(200).json({
                 success: true,
